move agg executors to `stream/executor/aggregation`
