+++ src/stream/src/executor/aggregation/agg_state.rs
@@ -24,9 +24,9 @@ use risingwave_pb::stream_plan::PbAggNodeVersion;
 use risingwave_storage::StateStore;
 
 use super::minput::MaterializedInputState;
-use super::{AggStateCacheStats, GroupKey};
 use crate::common::StateTableColumnMapping;
 use crate::common::table::state_table::StateTable;
+use crate::executor::aggregation::agg_group::{AggStateCacheStats, GroupKey};
 use crate::executor::{PkIndices, StreamExecutorResult};
 
 /// Represents the persistent storage of aggregation state.
