+++ src/stream/src/executor/aggregation/distinct.rs
@@ -21,8 +21,9 @@ use risingwave_common::array::{ArrayRef, Op};
 use risingwave_common::bitmap::{Bitmap, BitmapBuilder};
 use risingwave_common::row::{self, CompactedRow, RowExt};
 use risingwave_common::util::iter_util::ZipEqFast;
+use risingwave_expr::aggregate::AggCall;
 
-use super::{AggCall, GroupKey};
+use super::agg_group::GroupKey;
 use crate::cache::ManagedLruCache;
 use crate::common::metrics::MetricsInfo;
 use crate::executor::monitor::AggDistinctDedupMetrics;
