+++ src/stream/src/executor/aggregation/mod.rs
@@ -12,27 +12,73 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use agg_group::*;
-pub use agg_state::*;
-pub use distinct::*;
+use std::collections::HashMap;
+
 use risingwave_common::array::ArrayImpl::Bool;
 use risingwave_common::array::DataChunk;
 use risingwave_common::bail;
 use risingwave_common::bitmap::Bitmap;
 use risingwave_expr::aggregate::{AggCall, AggType, PbAggKind};
 use risingwave_expr::expr::{LogReport, NonStrictExpression};
-use risingwave_storage::StateStore;
+use risingwave_pb::stream_plan::PbAggNodeVersion;
 
-use crate::common::table::state_table::StateTable;
-use crate::executor::error::StreamExecutorResult;
+use crate::executor::prelude::*;
 
 mod agg_group;
 mod agg_state;
 mod agg_state_cache;
 mod distinct;
+mod hash_agg;
 mod minput;
+mod simple_agg;
+mod stateless_simple_agg;
 
-pub async fn agg_call_filter_res(
+pub use agg_state::AggStateStorage;
+pub use hash_agg::HashAggExecutor;
+pub use simple_agg::SimpleAggExecutor;
+pub use stateless_simple_agg::StatelessSimpleAggExecutor;
+
+/// Arguments needed to construct an `XxxAggExecutor`.
+pub struct AggExecutorArgs<S: StateStore, E: AggExecutorExtraArgs> {
+    pub version: PbAggNodeVersion,
+
+    // basic
+    pub input: Executor,
+    pub actor_ctx: ActorContextRef,
+    pub info: ExecutorInfo,
+
+    // system configs
+    pub extreme_cache_size: usize,
+
+    // agg common things
+    pub agg_calls: Vec<AggCall>,
+    pub row_count_index: usize,
+    pub storages: Vec<AggStateStorage<S>>,
+    pub intermediate_state_table: StateTable<S>,
+    pub distinct_dedup_tables: HashMap<usize, StateTable<S>>,
+    pub watermark_epoch: AtomicU64Ref,
+
+    // extra
+    pub extra: E,
+}
+
+pub trait AggExecutorExtraArgs {}
+
+pub struct SimpleAggExecutorExtraArgs {
+    pub must_output_per_barrier: bool,
+}
+impl AggExecutorExtraArgs for SimpleAggExecutorExtraArgs {}
+
+/// Extra arguments needed to construct an `HashAggExecutor`.
+pub struct HashAggExecutorExtraArgs {
+    pub group_key_indices: Vec<usize>,
+    pub chunk_size: usize,
+    pub max_dirty_groups_heap_size: usize,
+    pub emit_on_window_close: bool,
+}
+impl AggExecutorExtraArgs for HashAggExecutorExtraArgs {}
+
+async fn agg_call_filter_res(
     agg_call: &AggCall,
     chunk: &DataChunk,
 ) -> StreamExecutorResult<Bitmap> {
@@ -63,7 +109,7 @@ pub async fn agg_call_filter_res(
     Ok(vis)
 }
 
-pub fn iter_table_storage<S>(
+fn iter_table_storage<S>(
     state_storages: &mut [AggStateStorage<S>],
 ) -> impl Iterator<Item = &mut StateTable<S>>
 where
