+++ src/stream/src/from_proto/hash_agg.rs
@@ -26,8 +26,7 @@ use super::agg_common::{
 };
 use super::*;
 use crate::common::table::state_table::StateTable;
-use crate::executor::HashAggExecutor;
-use crate::executor::agg_common::{AggExecutorArgs, HashAggExecutorExtraArgs};
+use crate::executor::aggregation::{AggExecutorArgs, HashAggExecutor, HashAggExecutorExtraArgs};
 
 pub struct HashAggExecutorDispatcherArgs<S: StateStore> {
     args: AggExecutorArgs<S, HashAggExecutorExtraArgs>,
