+++ src/stream/src/from_proto/simple_agg.rs
@@ -22,8 +22,9 @@ use super::agg_common::{
 };
 use super::*;
 use crate::common::table::state_table::StateTable;
-use crate::executor::SimpleAggExecutor;
-use crate::executor::agg_common::{AggExecutorArgs, SimpleAggExecutorExtraArgs};
+use crate::executor::aggregation::{
+    AggExecutorArgs, SimpleAggExecutor, SimpleAggExecutorExtraArgs,
+};
 
 pub struct SimpleAggExecutorBuilder;
